Added Volume command and moved music commands into their own folder
