+++ AnotherTwitchChatBot Class Library/AnotherTwitchChatBot Class Library.csproj
@@ -131,13 +131,14 @@
     <Compile Include="Models\Commands\8BallCommand.cs" />
     <Compile Include="Models\Commands\Command.cs" />
     <Compile Include="Models\Commands\CommandFactory.cs" />
-    <Compile Include="Models\Commands\CurrentSongCommand.cs" />
+    <Compile Include="Models\Commands\Music\CurrentSongCommand.cs" />
     <Compile Include="Models\Commands\DoorCommand.cs" />
     <Compile Include="Models\Commands\GiveawayCommand.cs" />
-    <Compile Include="Models\Commands\PlayCommand.cs" />
-    <Compile Include="Models\Commands\PauseCommand.cs" />
-    <Compile Include="Models\Commands\SkipCommand.cs" />
-    <Compile Include="Models\Commands\SongRequestCommand.cs" />
+    <Compile Include="Models\Commands\Music\VolumeCommand.cs" />
+    <Compile Include="Models\Commands\Music\PlayCommand.cs" />
+    <Compile Include="Models\Commands\Music\PauseCommand.cs" />
+    <Compile Include="Models\Commands\Music\SkipCommand.cs" />
+    <Compile Include="Models\Commands\Music\SongRequestCommand.cs" />
     <Compile Include="Models\Giveaways\FollowerGiveaway.cs" />
     <Compile Include="Models\Giveaways\Giveaway.cs" />
     <Compile Include="Models\Giveaways\SubscriberGiveaway.cs" />
